Add pt-heartbeat start & execution script to mariadb
Imported heartbeat manifest from the old coredb_mysql module.
For now, it is not compatible with systemd. It should be enough
for testing for now (only a handful of db systems are on jessie).

This commit does not add pt-heartbeat to any server for now,
only its module. It should be tested extensively, specially
given that it is running now on the old coredb masters.

This should only be included on masters using the mariadb module.

Bug: T114752
Change-Id: Ifff701e3b8376f77a118c827b5d14ccaef6a3c46