Allow ssl key usage
If ssl is enabled, import keys and certificates from the private
repository and set them with restricted privileges on the subdir
/etc/mysql/ssl.

For now, we will share the certificates, probably we should
generate one per host in the future.

The certificates and keys have to exist first before deploying
this commit.

Change-Id: Ice0a9c81b2815cce99aa591bbac66ba75a8eb123
References: T111654