Don't set password in /root/.my.cnf if it is not defined
This will not change the default behavior as the default
value of $password in mariadb::config is 'undefined', which is
actually the string 'undefined', not undef (or false) in puppet/ruby.

On hosts where there is no root password, this will allow root shells
to connect to mysql with no password.  Currently it is passing
'undefined' as the password.

Bug: T127991
Change-Id: Ie75839a695386f536604a67127c9c81b456e3537