Make /opt/wmf-mariadb10/bin/mysqld_safe managed by puppet
This file is shipped by the custom wmf-mariadb10 package, but make it managed
via puppet to be able to quickly make changes without rebuilding the package.

This introduces the current, pristine version as shipped by wmf-mariadb10, to
make followup changes more transparent.

Note that /opt/wmf-mariadb10/bin/mysqld_safe on the servers is owned by
jynus:mysql, but this patch makes it owned by root:mysql (the ownership can
be cleaned up in a future update of wmf-mariadb10.

Thus patch doesn't add the class to any of the sub mariadb roles yet, this
can be done gradually with some non-core DBs first.

Bug: T145378
Change-Id: Ia95a3fa54eec64ae12743ed8bd114551413af8f3