mariadb:Create a systemd unit & init.d script for new package
The systemd unit has been on purpose not included on the package
so it can be handled by puppet. The config changes should also help
securize running mysqld.

The only configurable variables are for now @basedir and
@open_files_limit (that last one only on systemd).

Until we upgrade to 10.1 or 5.7 and get rid of trustys, we will
continue using init.d due to the lack of support on the older
(for us, current) versions for systemd.

Bug: T147305
Change-Id: I23cee27c8e86c94dc830cde799a50955f9513c9b