Overwrite (or create) init.d with custom one
Old package require manual installation. Now it is done
automatically.
New packages do not include it by default (or they are wrong).
This class overwrites the init.d script with the one that is
current installed from the old packages, to maintain 100%
compatibility.

In the future, mariadb::service will include a custom systemd
unit, but that is not yet possible with mariadb10.

This has to be parametrized to allow different packages other
than 10.0.

Change-Id: I3543a095bd240c190155aeb2753ba9605edc1b1f