mariadb: Add the posibility of selecting other package versions
The previous puppetization only allowed mariadb10 => true/false,
which was not really too future-proof.

This deprecates that paramenter (but continues to use it) and
introduces a new one called package. In the future we could also
add a version to hold packages to a particular existing/old
version for programmed upgrades.

As of now, versions allowed are wmf-mariadb10, wmf-mariadb101,
and wmf-mysql57 (maybe in the future the class can be renamed
to something else).

Bug: T149422
Change-Id: Ieb392eb1733dfa9df57107978a0170bb3f3b35a5