mariadb: Added calculation for gtid_domain_id
The gtid_domain_id variable needs to be unique and for now it should
be calculated the same way $server_id is, to make them independent
from each other

At the moment we are going to calculate the gtid_domain_id the same
way we calculate server_id. But having them independent from each other
will allow us to change one of them without affecting each other in the
future if for any reason we need to change the way they are calculated.
They both need to be unique across all the servers.

Bug: T149418
Change-Id: If60eba4597ccb7ce2fbd4bdafa4ad647bc3ab8e7