Install wmf-mariadb-client for client-only installs
The generic mariadb-client or mysql-client does not support
openssl-based TLS properly, and errors-out when tried to use
with a wmf-compiled mariadb server package. Instead of using
the upstream package, use a customly-compiled binary.

It will generate more load for maintaining-but it takes very
little extra time compared when compiling the server, and
it normally is not a source of security issues, because it
only contains a few small client binaries and light libraries
imeplementing the mysql client protocol.

We use the Mariadb 10.1 everywhere because it will be compatible
with almost all mysql and mariadb versions out there, while
providing the best support and complete features.

Bug: T157702
Change-Id: I49e7b9e5b65942fbfe2101943ad23fed7c755334