Remove old CA (ssl='on') and add a new option "socket"
There are no longer servers using the temporary, 1-year old
CA- remove it but maintain the option ssl='on' by redefining it
as an alias to 'puppet-cert'.

Add a new config option socket- this has to be supported by
the config template, so on deploy this will be mostly a noop.

Bug: T157702
Bug: T143896
Bug: T148507
Change-Id: I2c2b01dca929351911b1b4780a132e487cf2a7b5